Move keri event parsing tests to `cesr_adapter`
